vastspace: Fix Island3Entity won't raise exception by just selecting
Island3Entity is not creatable from scripts directly; it's automatically
created when an Island3 (a CoordSys subclass) is created.  In old times,
we had only way to prevent scripts from creating such Entity, i.e.
not defining EntityRegister.  But lack of EntityRegister cause errors on
scripts when just selecting such an Entity.

Now that we have EntityRegisterNC for that purpose, we can prevent both
creation and errors in scripts since 5a6a574e7.