Redefine absolute magnitude of an Astrobj to brightness
We had been using absolute magnitude for celestial bodies such
as stars, but it turned out that it has undesirable issues.

Absolute magnitude is useful in that it can express brightnesses in
recognizable scale even if dynamic range of brightnesses are wide,
but it's only for humans.  Computers do not care how many zeros are
lined up.  Even worse, converting to and from logarithmic scale
all the time when performing computation is waste of CPU time
and looses significant digits in floating point numbers.

There have been another issue in sunlight reflection calculation.
Reflected light was sometimes stronger than direct sunlight, which
isn't realistic.  I haven't identified exact cause of this issue,
but it would be gone once we quit using logarithmic scale and do
unnecessary computations.

Now we use absolute brightness (in linear scale) internally and calculate
absolute (or apparent) magnitude only when necessary.  The scale of
brightness is set so that the Sun's brightness is 1.