surface: Remove orientation stabilizer for Apache when controlled by Player
The stabilizer is introduced to compensate instability of old algorithm,
but now the algorithm itself is improved that the stabilizer is not as
useful.

Now that the stabilizer does not get in the way, you can perform
acrobatic flights such as upside down.

AI controlled Apaches continue to use the stabilizer.