Revive the Earth's detail textured version of GLSL shader
Earth.fs and earth.vs have been broken since DrawTextureCubeEx is used to
render the Earth.  It's because assumed coordinate systems are different
and we could not share single shader for both DrawTextureSphere and
DrawTextureCubeEx.  Now that DrawTextureCubeEx seems to be accepted as
the next generation of planet rendering algorithm, we just throw away
DrawTextureSphere version and adapt to DrawTextureCubeEx.

DrawTextureSpheroid uses the same coordiante system convention as the
latter, so we can use the shader with it.