Move GLwindow system's transformation matrix setup to make the invocation only once
Previously, both the projection and modelview matrices are set up in each call
to GLwindow::drawInt(), which is not optimal since the transformations are the
same among GLwindows.  This construct had an advantage that even if some
GLwindow instances tamper with the transformation, that won't affect other
windows, but the bottom line is that every GLwindow should take care of their
rendering routines and they should push the matrices by their own if necessary.

Actually, it was not convenient when we override GLwindow::drawInt() to make a
frameless window.  We had to repeat writing transformations code in the
overridden drawInt() like the base class.

Now, the transformations are performed only once in the startup of rendering of
GLwindows.