Add mouse positions information to mouse click event for GLwindowState
GLwindowState had mx and my fields that are used for mouse coordinates in
mouse move events, but they are not set in mouse down or up event.

These fields are now explicitly set in these events, too.

There are also complications in mousex and mousey fields, but it's another
story.