Make items in the inventory window drag & droppable to another window
The items can be moved between ships and stations' cargo holds by the GUI.

It should check the distance and connectivity between ships to determine whether
the items are transferable, but currently items can be moved in instant no matter
how distant the sender and the receiver.

This user interface is similar to Windows Explorer and players of Eve Online
or Space Engineers won't find difficulties.