Add SSD option to normalize planet texture colors, demonstrated with Earth
The Earth texture has low brightness pixels, which makes the landscape
around the pixel too dark.  This is not realistic, so the pixels are
normalized as 3-component vector to make the brightness consistent.

Also replaced RoundAstrobj::texname with Texture object, since we want to
treat the texture as the same as extextures.