Introduce DrawTextureSphere-forward.h to clarify dependencies
RoundAstrobj class itself should not know about DrawTextureSphere since
it can be used in the server, where no graphics are involved.
However, some of the data structures are necessary for even the server,
such as TerrainMods.

We had defined these shared structures inside RoundAstrobj to resolve
this issue, but it's not really desirable to contain definition about
another structure's information.

Adding forward declaration header cleanly resolves this mess, in
expense of file count.  But I believe it deserves.