Make the base LOD of DrawTextureCubeEx single tile
Previously, we have divided the LOD 0 patches to 4x4 subpatches,
result in 64x64 squares in one side of the cube, which is too detailed
when they are hidden behind horizon.
So we just use a single tile in LOD 0 and let the LOD tree adapt to
optimal grid subdivision.

Essentially, the LOD definition is shifted by 1.