Modify specification of SSD's texture parameters
The cloudSync parameter was specified as the fourth parameter of extexture
command in SSDs.  This positional parameter specification is so obscure
that no one would recognize what the 0 or 1 means!  Not to mention there
is no proper documentation.

So I removed the fourth positional parameter and replaced with named
parameter, just like alpha or height flags.  This way the command looks
much easier to understand.

The cloud shadow itself is not rendered by shader issues, so we are still
unable confirm it works.