Fix DrawTextureCubeEx wouldn't determine terrain LODs accurately
It didn't consider the height variation itself (always determined the
level necessary for sea level).  So it could awfully less detailed when
on top of a high mountain or bottom of deep under-sea level valley.

Now, the LOD determination is done considering local terrain height
(which is most precise), but for LODs only higher than 3.
This is to reduce calls to getTerrainHeight() necessary every frame.