Allow multiple LODs cover the same z-slice as WarSpace
Since maximum LOD number has been increased significantly for smooth
terrain, rendering only the most detailed LOD for WarSpace is not enough;
you can occasionally see other Entities through terrain.

So we introduce zBufLODs parameter to determine which LODs should
be rendered in the same z-slice as WarSpace.  This parameter is defined as
count of LODs from the most detailed LOD, in other words, actual LODs for
rendering with WarSpace is between (maxLODs - zBufLODs) and (maxLODs).
This design is to enable tweaking maximum LOD without modifying zBufLODs
as well.

Also, LODs that would be rendered in later z-slice are skipped in former
z-slice to prevent duplicate rendering.  This could make holes on planets
when you're close to the celestial body and not in a SurfaceWar.

Also aggregated terrain noise parameters to a structure TerrainNoise to
handle them in a uniform way.  And it should be harder to forget some
coding when adding another parameter if there are fewer repeating codes.

There's an issue that the cloud is overwritten by nearer z-slices.
This issue is related to the nature of z-slicing and could be hard to
address.  Probably we could make clouds a set of blobs instead of
continuous surface, which is hard to distribute among z-slices.