Add terrainTextureHeight SSD entry for specifying height of height map
Previously, height of height map is determined from terrainNoiseHeight,
which is not necessarily the same.  In particular, the moon has high
variation of heights in the height texture, but height noises should be
lower than that.

Also, height is sampled for each vertex in the patch mesh, rather than
only the corner points.  This may require some extra computation cost,
but it seems negligible even in debug mode.