Reduce calls to get viewport for improving performance
According to VC2015's performance profiler, the call to glGetIntegerv
is not insignificant.  We have used it every frame, but we really only
need to call it when the size of the window changes.  So we call it on
reshape_func() (which is really an event handler for WM_SIZE) and cache
the result for use in display_func().