Fix some loopholes in variable substitution in SSDs
Previously, the variable expansion only happened to the variable
substitution ($blah) at the beginning of a string or anywhere in a
double-quoted string.
Now you can place dollar sign in any place of any command.

We came to the conclusion that we can't really replace "define" command
with scripted procedures like the one below because the string evaluation
forces the type of variables to be strings, while Squirrel expressions
do not allow automatic conversion from a string to a number.

	proc define {id val} {
		upvar $id myid
		set myid [expr ($val).tofloat()]
	}

So we cannot just use "set" for everything and use Squirrel expressions
in "expr" command at the same time.  Our goal is not to mimic the syntax
of Tcl, but to create an easy way to write specification, so we keep the
native implementation of "define".

The rule of thumb is:

  * Use "set" for string substitutions.
  * Use "define" for numerical constants.